Update
So, this is where I've decided to... not throw in the towel, necessarily, but maybe hang it up for a bit. I think my background in programming has not prepared me quite enough for this assessment. While I'm confident that I could learn this with some guidance, I don't have the tools at the moment to go forward on my own. I'd love to get your perspective on how to proceed!